+++ apps/artemis-web/components/ongoing-hijack-table/ongoing-hijack-table.tsx
@@ -561,7 +561,7 @@ const OngoingHijackTableComponent = (props: any): ReactElement => {
                 <SizePerPageDropdownStandalone {...paginationProps} />
               </div>
               <div style={{ float: "right" }}>
-                <PaginationTotalStandalone {...paginationProps} />
+                {/* <PaginationTotalStandalone {...paginationProps} /> */}
                 <PaginationListStandalone {...paginationProps} />
               </div>
             </div>
@@ -595,6 +595,16 @@ const OngoingHijackTableComponent = (props: any): ReactElement => {
               {...toolkitprops.baseProps}
               {...paginationTableProps}
             />
+            <div className="row">
+              <div className="col-lg-12">
+                <div style={{ float: "right" }}>
+                  <PaginationListStandalone {...paginationProps} />
+                </div>
+                <div style={{ float: "left" }}>
+                  <PaginationTotalStandalone {...paginationProps} />
+                </div>
+              </div>
+            </div>
           </>
         );
       }}